send_once behavior for resque.
This adds the ability to call Job.create_once or Resque.enqueue_once
to add jobs to Resque. When using this syntax, the job will only be
queued up if the job is not already in the queue. This is nice if you
have an expensive process that should only run once, but might be
triggered simultaneously by several different system events.